# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bus Reservation System API
  description: |
    Comprehensive API documentation for the Bus Reservation System. 
    Includes:
    1. Public Search API (Searching buses, filtering, sorting).
    2. Operator API (Manifests, bookings, passenger lists).
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
paths:
...
  /admin/add-operator:
    post:
      summary: Create a new Operator (Admin Only)
      description: |
        Registers a new user with the role of 'operator'. 
        Requires a valid JWT token with 'admin' role.
      operationId: addOperator
      tags:
        - Admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest' 
      responses:
        '201':
          description: Operator created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - User is not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - Email Already Exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Email Already Exists
        '500':
          $ref: '#/components/responses/InternalServerError'

  
  # --- Operator Managing Routes and seats ---
...
  /operator/routes:
    post:
      summary: Create a New Route (Operator Only)
      description: |
        Creates a route (Source -> Destination) and optionally defines intermediate stops.
        This route can later be assigned to specific bus trips.
      operationId: createRoute
      tags:
        - Operator
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
      responses:
        '201':
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '403':
          description: Forbidden - Only Operators can create routes
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      summary: Get All My Routes
      description: Returns a list of all routes created by the logged-in operator.
      operationId: getMyRoutes
      tags:
        - Operator
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of routes
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        route_id:
                          type: integer
                        source:
                          type: string
                        destination:
                          type: string
                        total_distance:
                          type: number
...
  /admin/operators:
    get:
      summary: Get all Operators
      description: Returns a list of all registered operators. Useful for populating dropdowns.
      operationId: getOperators
      tags:
        - Admin
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of operators
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        user_id:
                          type: integer
                          example: 5
                        name:
                          type: string
                          example: "KPN Travels"
                        email:
                          type: string
                          example: "kpn@example.com"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'                   
...
  /trips/{trip_id}/layout:
    get:
      summary: Get Seat Layout for a Trip
      description: Returns the bus layout with real-time availability. Passes search context for dynamic pricing.
      operationId: getTripLayout
      tags:
        - Public Search
      parameters:
        - in: path
          name: trip_id
          schema:
            type: integer
          required: true
          description: The ID of the Trip (not the Bus ID)
        - in: query
          name: pickup_id
          schema:
            type: integer
          required: false
          description: ID of the pickup stop (to calculate exact fare)
        - in: query
          name: drop_id
          schema:
            type: integer
          required: false
          description: ID of the drop stop (to calculate exact fare)
        
        # --- NEW PARAMETERS START ---
        - in: query
          name: from_city
          schema:
            type: string
          required: false
          description: Origin city name from search (e.g. "coimbatore") to track demand correctly.
        - in: query
          name: to_city
          schema:
            type: string
          required: false
          description: Destination city name from search (e.g. "bangalore") to track demand correctly.
        # --- NEW PARAMETERS END ---

      responses:
        '200':
          description: Layout retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripLayoutResponse'
        '404':
          description: Trip not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /bookings:
    post:
      summary: Booking seats for a trip
      description: |
        Receives a list passengers and seat selections, checks availability (freezing logic),
        and creates a confirmed booking.
      operationId: createBooking
      tags:
        - Booking Flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookingRequest'
      responses:
        '201':
          description: Successful booking response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Booking created successfully"
                  booking_id:
                    type: integer
                    example: 9001
                  ticket_id:
                    type: integer
                    example: 9190
                  total_amount:
                    type: integer
                    example: 1000.00
                  status:
                    type: string
                    example: "Confirmed"
                  
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          description: Conflict - Seats already taken
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Seat Unavailable"
                  message:
                    type: string
                    example: "Seat L1 is already booked by another user."
        '500':
          $ref: '#/components/responses/InternalServerError'  
...
  /getBookingsInfo:
    get:
      summary: Get Operator Manifest
      description: |
        Returns the full manifest for a specific Operator.
        Data is grouped by Bus -> Trips -> Passengers.
      operationId: getBookingsInfo
      tags:
        - Operator Info
      parameters:
        - name: operator_id
          in: query
          required: true
          schema:
            type: integer
            example: 1
          description: The ID of the operator requesting the manifest.

      responses:
        '200':
          description: Successful manifest response.
          content:
            application/json:
              schema:
                type: object
                properties:
                  bus:
                    type: array
                    description: List of buses found for the operator.
                    items:
                      $ref: '#/components/schemas/OperatorBusManifest'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SignupRequest:
      type: object
      required: [name, email, password, phone]
      properties:
        name:
          type: string
          example: "Alice"
        email:
          type: string
          format: email
          example: "alice@example.com"
        password:
          type: string
          format: password
          example: "secret123"
        phone:
          type: string
          example: "9876543210"
...
    RouteStopRequest:
      type: object
      required:
        - stop_name
        - order_id
        - stop_type
      properties:
        stop_name:
          type: string
          example: "Vellore"
        order_id:
          type: integer
          example: 1
        distance:
          type: number
          description: Distance from source
          example: 140.5
        price:
          type: number
          description: Price from source to this stop
          example: 300.00
        estimated_time:
          type: string
          description: E.g., "2 hrs 30 mins"
          example: "2h 30m"
        stop_type:
          type: string
          enum: ['Boarding', 'Dropping', 'Both']
          example: "Boarding"
    RouteRequest:
      type: object
      required:
        - route_name
        - total_distance
        - estimated_time
        - stops
      properties:
        route_name:
          type: string
          description: E.g., "Chennai - Bangalore"
          example: "Chennai - Bangalore"
        total_distance:
          type: number
          example: 350.00
        estimated_time:
          type: string
          example: "6h 00m"
        stops:
          type: array
          items:
            $ref: '#/components/schemas/RouteStopRequest'
    RouteResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Route created successfully with 2 stops"
        route_id:
          type: integer
          example: 201

    # --- Trip Management ---
...
    SeatLayoutPassenger:
      type: object
      properties:
        seat_id:
          type: integer
          example: 15
        seat_number:
          type: string
          example: "U5"
        seat_type:
          type: string
          example: "Sleeper"
        price:
          type: number
          example: 850.00
        status:
          type: string
          enum: ['AVAILABLE', 'BOOKED']
          example: "AVAILABLE"
    TripLayoutResponse:
      type: object
      properties:
        trip_id:
          type: integer
        bus_id:
          type: integer
        bus_type:
          type: string
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatLayoutPassenger'
        boarding_points:
          type: array
          items:
            type: object
            properties:
              stop_id:
                type: integer
              stop_name:
                type: string
              time:
                type: string
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login Successful"
        token:
          type: string
          description: "The JWT Token to use for future requests"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            user_id:
              type: integer
              example: 15
            name:
              type: string
              example: "Alice"
            email:
              type: string
              example: "alice@example.com"
    # --- New Schema for Search Results ---
...
    BookingRequest:
      type: object
      required: [user_id, trip_id, pickup_stop_id, drop_stop_id, passengers]
      properties:
        user_id:
          type: integer
          description: Id of the user making the booking
          example : 101
        trip_id:
          type: integer
          description: Id of the trip they are booking
          example : 505
        pickup_stop_id:
          type: integer
          description: Id of the pickup stop
          example : 10
        drop_stop_id:
          type: integer
          description: Id of the drop stop
          example : 20
        passengers:
          type: array
          description: List of passengers on this bus and their assigned seats.
          minItems: 1
          items:
            type: object
            required: [name, age, gender, seat_number, seat_id]
            properties:
              name:
                type: string
                example: "Rajesh Kumar"
              age:
                type: integer
                example: 25
              gender:
                type: string
                enum: ["Male", "Female", "Other"]
                example: "Male"
              seat_number:
                type: string
                description: Visual Seat number 
                example: "A1"
              seat_id:
                type: integer
                description: Database Id of the seat
                example: 101


    # --- Saved Travellers ---
...
    OperatorBusManifest:
      type: object
      properties:
        bus_type:
          type: string
          example: "AC Sleeper"
        bus_number:
          type: string
          example: "TN-01-AA-1001"
        passengers_list:
          type: array
          description: List of passengers on this bus.
          items:
            type: object
            properties:
              trip_id:
                type: integer
                example: 5001
              ticket_id:
                type: integer
                example: 9001
              booking_id:
                type: integer
                example: 1001
              passenger_name:
                type: string
                example: "Rajesh Kumar"
              seat_number:
                type: string
                example: "A1"
              seat_condition:
                type: string
                enum: [good, repair, maintenance, damaged]
                example: "good"
              contact_number:
                type: string
                example: "9876543210"
              route_name:
                type: string
                example: "Chennai - Madurai"
              pickup_point:
                type: string
                example: "Chennai Koyambedu"
              drop_point:
                type: string
                example: "Madurai Mattuthavani"
              departure_time:
                type: string
                format: date-time
                example: "2025-12-10 21:00:00.000000"
              booking_date:
                type: string
                format: date-time
                example: "2025-12-01 10:00:00.000000"
              booking_status:
                type: string
                example: "Confirmed"
              amount_paid:
                type: number
                format: float
                example: 600.0
              total_trip_bookings:
                type: integer
                description: Total passengers on this specific trip.
                example: 3

    

    # --- Shared Error Schemas ---
    ErrorResponse:
      type: object
      description: Standard error wrapper for all non-2xx responses.
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Invalid date format."
        code:
          type: string
          example: "VAL_001"
      required: [status, error, message]
  responses:
    BadRequestError:
      description: Bad Request – Invalid inputs (e.g., date format or missing ID).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized – missing or invalid authentication credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden – authenticated but not allowed to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error – unexpected condition on the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'